+++ jobbergate-api/jobbergate_api/apps/job_scripts/routers.py
@@ -19,6 +19,7 @@ from jobbergate_api.apps.job_scripts.schemas import (
     JobScriptCloneRequest,
     JobScriptCreateRequest,
     JobScriptDetailedView,
+    JobScriptFileDetailedView,
     JobScriptListView,
     JobScriptUpdateRequest,
     RenderFromTemplateRequest,
@@ -282,8 +283,7 @@ async def job_script_get_file(
     Note:
         See https://fastapi.tiangolo.com/advanced/custom-response/#streamingresponse
     """
-    job_script = await secure_services.crud.job_script.get(id)
-    job_script_file = await secure_services.file.job_script.get(job_script.id, file_name)
+    job_script_file = await secure_services.file.job_script.get(id, file_name)
     return StreamingResponse(
         content=await secure_services.file.job_script.stream_file_content(job_script_file),
         media_type="text/plain",
@@ -294,24 +294,36 @@ async def job_script_get_file(
 @router.put(
     "/{id}/upload/{file_type}",
     status_code=status.HTTP_200_OK,
-    description="Endpoint to upload a file to a job script file",
+    description=(
+        "Endpoint to upload a file to a job script. "
+        "If a previous filename is provided, the file will be renamed from that. "
+        "Upload file is optional in this scenario since the file content can be copied from previous file."
+    ),
+    response_model=JobScriptFileDetailedView,
 )
 async def job_script_upload_file(
     id: int = Path(...),
     file_type: FileType = Path(...),
-    upload_file: UploadFile = File(..., description="File to upload"),
+    filename: str | None = Query(None, max_length=255),
+    upload_file: UploadFile | None = File(None, description="File to upload"),
+    previous_filename: str | None = Query(
+        None, description="Previous name of the file in case a rename is needed", max_length=255
+    ),
     secure_services: SecureService = Depends(
         secure_services(Permissions.JOB_SCRIPTS_CREATE, ensure_email=True)
     ),
 ):
     """Upload a file to a job script."""
-    if upload_file.filename is None:
+    # This is included for backwards compatibility with the previous implementation
+    # where filename was recovered from the upload_file object
+    filename = filename or getattr(upload_file, "filename")
+    if not filename:
         raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="The upload file has no filename",
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Filename must be provided either as a query parameter or as part of the file upload",
         )
 
-    logger.debug(f"Uploading file {upload_file.filename} to job script {id=}")
+    logger.debug(f"Uploading file {filename=} to job script {id=} with {file_type=} and {previous_filename=}")
 
     job_script = await secure_services.crud.job_script.get(id)
     if Permissions.ADMIN not in secure_services.identity_payload.permissions:
@@ -319,10 +331,11 @@ async def job_script_upload_file(
             job_script, owner_email=secure_services.identity_payload.email
         )
 
-    await secure_services.file.job_script.upsert(
+    return await secure_services.file.job_script.upsert(
         parent_id=job_script.id,
-        filename=upload_file.filename,
+        filename=filename,
         upload_content=upload_file,
+        previous_filename=previous_filename,
         file_type=file_type,
     )